Move python equation of states and modules to build_wheel
