+++ build_wheel/Cargo.toml
@@ -9,6 +9,8 @@ edition = "2018"
 crate-type = ["cdylib"]
 
 [dependencies]
+quantity = "0.5"
 feos-core = { path = "..", features = ["python"] }
 pyo3 = { version = "0.16", features = ["extension-module", "abi3", "abi3-py37"] }
+numpy = "0.16"
