+++ src/python/cubic.rs
@@ -1,15 +1,11 @@
-use crate::cubic::{PengRobinson, PengRobinsonParameters, PengRobinsonRecord};
+use crate::cubic::{PengRobinsonParameters, PengRobinsonRecord};
 use crate::joback::JobackRecord;
 use crate::parameter::{IdentifierOption, Parameter, ParameterError, PureRecord};
 use crate::python::joback::PyJobackRecord;
-use crate::python::parameter::{PyBinaryRecord, PyChemicalRecord, PyIdentifier};
+use crate::python::parameter::PyIdentifier;
 use crate::*;
-use numpy::convert::ToPyArray;
-use numpy::{PyArray1, PyArray2};
-use pyo3::exceptions::PyValueError;
+use numpy::PyArray2;
 use pyo3::prelude::*;
-use quantity::python::*;
-use std::collections::HashMap;
 use std::convert::TryFrom;
 use std::rc::Rc;
 
@@ -71,64 +67,4 @@ impl PyPengRobinsonParameters {
     fn __repr__(&self) -> PyResult<String> {
         Ok(self.0.to_string())
     }
-}
-
-/* EQUATION OF STATE */
-
-/// A simple version of the Peng-Robinson equation of state.
-///
-/// Parameters
-/// ----------
-/// parameters : PengRobinsonParameters
-///     The parameters of the Peng-Robinson equation of state to use.
-///
-/// Returns
-/// -------
-/// PengRobinson
-#[pyclass(name = "PengRobinson", unsendable)]
-#[pyo3(text_signature = "(parameters)")]
-#[derive(Clone)]
-pub struct PyPengRobinson(pub Rc<PengRobinson>);
-
-#[pymethods]
-impl PyPengRobinson {
-    #[new]
-    fn new(parameters: PyPengRobinsonParameters) -> Self {
-        Self(Rc::new(PengRobinson::new(parameters.0.clone())))
-    }
-}
-
-impl_equation_of_state!(PyPengRobinson);
-impl_virial_coefficients!(PyPengRobinson);
-
-impl_state!(PengRobinson, PyPengRobinson);
-impl_state_molarweight!(PengRobinson, PyPengRobinson);
-impl_vle_state!(PengRobinson, PyPengRobinson);
-// impl_estimator!(PengRobinson, PyPengRobinson);
-
-#[pymodule]
-pub fn cubic(_py: Python<'_>, m: &PyModule) -> PyResult<()> {
-    m.add_class::<PyIdentifier>()?;
-    m.add_class::<Verbosity>()?;
-    m.add_class::<Contributions>()?;
-    m.add_class::<PyChemicalRecord>()?;
-    m.add_class::<PyJobackRecord>()?;
-
-    m.add_class::<PyPengRobinsonRecord>()?;
-    m.add_class::<PyPureRecord>()?;
-    m.add_class::<PyBinaryRecord>()?;
-    m.add_class::<PyPengRobinsonParameters>()?;
-
-    m.add_class::<PyPengRobinson>()?;
-    m.add_class::<PyState>()?;
-    m.add_class::<PyPhaseDiagramPure>()?;
-    m.add_class::<PyPhaseDiagramBinary>()?;
-    m.add_class::<PyPhaseDiagramHetero>()?;
-    m.add_class::<PyPhaseEquilibrium>()?;
-
-    // let utils = PyModule::new(py, "utils")?;
-    // utils.add_class::<PyDataSet>()?;
-    // utils.add_class::<PyEstimator>()?;
-    // m.add_submodule(utils)?;
-    Ok(())
 }
