+++ src/python/user_defined.rs
@@ -3,42 +3,10 @@ use crate::*;
 use ndarray::prelude::*;
 use num_dual::python::{PyDual3Dual64, PyDual3_64, PyDual64, PyHyperDual64, PyHyperDualDual64};
 use num_dual::{Dual, Dual3, Dual3_64, Dual64, DualVec64, HyperDual, HyperDual64};
-use numpy::convert::{IntoPyArray, ToPyArray};
-use numpy::{PyArray1, PyArray2};
-use pyo3::exceptions::PyValueError;
+use numpy::convert::IntoPyArray;
 use pyo3::prelude::*;
-use quantity::python::{PySIArray1, PySIArray2, PySINumber};
-use std::collections::HashMap;
+use quantity::python::PySIArray1;
 use std::fmt;
-use std::rc::Rc;
-
-/// Equation of state implemented as python class.
-///
-/// Parameters
-/// ----------
-/// obj : python class object
-///     The class that implements the equation of state.
-///
-/// Returns
-/// -------
-/// UserDefinedEos
-///
-/// Raises
-/// ------
-/// RunTimeError
-///     If the class does not implement all necessary methods.
-#[pyclass(name = "UserDefinedEos", unsendable)]
-#[derive(Clone)]
-#[pyo3(text_signature = "(obj)")]
-pub struct PyUserDefinedEos(Rc<PyEoSObj>);
-
-#[pymethods]
-impl PyUserDefinedEos {
-    #[new]
-    fn new(obj: Py<PyAny>) -> PyResult<Self> {
-        Ok(Self(Rc::new(PyEoSObj::new(obj)?)))
-    }
-}
 
 struct PyHelmholtzEnergy(Py<PyAny>);
 
@@ -252,34 +220,3 @@ impl_helmholtz_energy!(PyStateD3D, PyDual3Dual64, Dual3<Dual64, f64>);
 impl_helmholtz_energy!(PyStateD3DV2, PyDual3DualVec64_2, Dual3<DualVec64<2>, f64>);
 impl_helmholtz_energy!(PyStateD3DV3, PyDual3DualVec64_3, Dual3<DualVec64<3>, f64>);
 impl_helmholtz_energy!(PyStateF, f64, f64);
-
-impl_equation_of_state!(PyUserDefinedEos);
-impl_virial_coefficients!(PyUserDefinedEos);
-impl_state!(PyEoSObj, PyUserDefinedEos);
-impl_state_molarweight!(PyEoSObj, PyUserDefinedEos);
-impl_vle_state!(PyEoSObj, PyUserDefinedEos);
-
-#[pymodule]
-pub fn user_defined(_py: Python, m: &PyModule) -> PyResult<()> {
-    m.add_class::<PyStateHD>()?;
-    m.add_class::<PyStateD>()?;
-    m.add_class::<PyStateDDV3>()?;
-    m.add_class::<PyStateD3>()?;
-    m.add_class::<PyStateF>()?;
-    m.add_class::<PyStateHDD>()?;
-    m.add_class::<PyStateHDDV2>()?;
-    m.add_class::<PyStateHDDV3>()?;
-    m.add_class::<PyStateD3D>()?;
-    m.add_class::<PyStateD3DV2>()?;
-    m.add_class::<PyStateD3DV3>()?;
-    m.add_class::<Verbosity>()?;
-    m.add_class::<Contributions>()?;
-    m.add_class::<PyUserDefinedEos>()?;
-    m.add_class::<PyState>()?;
-    m.add_class::<PyPhaseEquilibrium>()?;
-    m.add_class::<PyPhaseDiagramPure>()?;
-    m.add_class::<PyPhaseDiagramBinary>()?;
-    m.add_class::<PyPhaseDiagramHetero>()?;
-    m.add_class::<PyPhaseEquilibrium>()?;
-    Ok(())
-}
